# services/backend-go/api/openapi.yaml
openapi: 3.1.0
info:
  title: SkoLab API
  version: "1.0"
  description: |
    The public API of skolab-api: the Go gateway. (The Python service runs in
    the same container and is not reachable from outside; compiles reach it
    through POST /api/v1/colab/compile.)

    Every error is `{"error": "<message for people>", "code": "<stable code>"}`;
    clients branch on `code`. Every response carries HSTS, nosniff and frame
    and referrer policies, and responses under /api are never cached, except
    the template catalog, which a browser may keep but must revalidate (ETag).

    The collaboration WebSocket (`GET /ws/colab/{workspace_id}?ticket=...`)
    is not described here: open it with a ticket from
    `POST /api/v1/ws/colab/{workspace_id}/tickets`.

    This file is the contract the QA tools test against (services/qa).
servers:
  - url: https://skolab-api.onrender.com
security:
  - firebase: []
...
paths:
...
  /api/v1/colab/compile:
    post:
      tags: [collaboration]
      operationId: compile
      summary: Compile LaTeX to PDF (charges the caller's quota; one at a time per user)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [latex_source]
              properties:
                latex_source: { type: string, minLength: 1, maxLength: 100000 }
                engine: { type: string, enum: [pdflatex] }
      responses:
        "200":
          description: The compile ran; `status` says whether it produced a PDF.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CompileResult" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "413": { $ref: "#/components/responses/TooLarge" }
        "429": { $ref: "#/components/responses/TooMany" }
        "502":
          description: The compile backend failed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503": { $ref: "#/components/responses/Unavailable" }
components:
  securitySchemes:
    firebase:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID token
    sentrySignature:
      type: apiKey
      in: header
      name: Sentry-Hook-Signature
...
  schemas:
    Error:
      type: object
      required: [error, code]
      additionalProperties: false
      properties:
        error: { type: string, description: For people; wording may change. }
        code: { type: string, pattern: "^[a-z_]+$", description: Stable; branch on this. }
...
    CompileResult:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [compiled, error, timeout] }
        pdf_base64: { type: [string, "null"] }
        log: { type: string }
        errors: { type: array, items: { type: string } }
...
  responses:
    BadRequest:
      description: The request is malformed or outside the contract.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Unauthorized:
      description: Missing, invalid, expired or revoked Firebase ID token.
      headers:
        WWW-Authenticate: { schema: { type: string } }
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Forbidden:
      description: Signed in, but not allowed (role, account policy, or someone else's resource).
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    TooLarge:
      description: Request body over 1 MiB (12 MiB for uploads and imports), or a project or file over its limit.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    TooMany:
      description: Rate limited or quota spent; see Retry-After.
      headers:
        Retry-After: { schema: { type: integer } }
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    Unavailable:
      description: A dependency (database, Firebase, compile backend) is unavailable.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }